requirements.txt
# HTTP & scraping
requests>=2.31,<3
aiohttp>=3.9,<4
beautifulsoup4>=4.12,<5
fastapi
uvicorn

# PDF Parsing
#unstructured[pdf]>=0.12
pdfplumber>=0.10
pymupdf>=1.23

# NLP & Embeddings
#sentence-transformers>=2.2
#transformers>=4.40
#torch>=2.2

# Vector DB
qdrant-client>=1.7

# Database
psycopg2-binary>=2.9
sqlalchemy>=2.0
alembic>=1.13

# Orchestration
langchain>=0.2
langchain-community>=0.2
langchain-openai>=0.1
langchain-text-splitters>=0.2

# Validation & Config
pydantic>=2.6
python-dotenv>=1.0

# Logging
loguru>=0.7